requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2024.1
Django==4.2.9
django-crispy-forms==2.2
django-filter==24.3
djangorestframework==3.15.2
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
